chore: delete `hardfork.rs` and co-locate trait def with impl
